Copy necessary parts of Noweb 2 and build them.
This is a direct copy of the C source files from Noweb2 (2.11b, I believe).
In this case we are building markup.exe and nt.exe.

We also need nocond and something to tie the input and a file together.
Unfortunately, for developers on Windows, nocond is a script and tie is not a standard command.

Nocond was built as a simple parser (and yes, I know that a massive switch statement is bad).
Tie was hacked together in even less time.
These are not designed to be general purpose tools, but one shot tools to enable the building of Noweb 3.

I imagine at some future point there will either be a better bootstrapping system or Noweb 3 will use the same trick Noweb 2 used --- that is distribute sources as well.