Build no and nwmtime.
This change builds no.exe and nwmtime.exe.

xpipe.nw is the largest change.
This is addition of support for the Windows platform.
This code worked in my experiments over the summer, and with this commit I am assuming that it still will.
I will verify when I get the lua scripts building.

nwtime.nw has had non-Posix versions added.
This is something in the github version that was not in the .tar.gz from the Noweb website.

nwmtime.nw included unistd.h, but it is not be needed so I simply removed it.

The change to Makefile was because of the realization that I had broken the Linux build in my previous two commits.
nmake doesn't work very well if your local variable is $(BIN), so it was replaced with $(BINDIR).
The `Makefile's which were using the BIN macro have been updated to use the BINDIR macro instead.