Adding support for nmake clean and clobber.
Updated almost all of the Makefile.win files to support a clean and clobber setting.
For these files I define clean as ``just the files I want are left'' and clobber as ``it's all gone''.
Probably the most controversial thing about this is that the root Makefile.win will ask for a clobber of some items when it is just being cleaned.
This fits with my definitions because even though a library may be clean by removing everything except the .lib files;
a system dependent on those .lib files isn't clean until the .lib files are clobbered.

I also added a large README.win to explain how to build, clean, and clobber everything on a Windows box.