Weave itself.
Addition of the tex directory from Noweb 2 to generate the nwmac.tex for TeX and noweb.sty for LaTeX.

Addition to Makefile.win to generate doc.pdf.
This only works with Noweb 3 and weaves everything, so this servers as some level of sanity testing.

Both READMEs updated because of the added directory.
README.win updated with instructions for making doc.pdf as well as timing experiments.