Get noweb 3 to build on linux and final merge
Changes to Makefile to build on linux.

Merge of README.win into README and corrections to match the
new build system.

Update of limake to handing the BIN->BINDIR rename.

Re-merge of the updated nwconfig.unix and nwconfig.win to be closer
to the original.

Install used to have duplicate build information which varied slightly
from what was in README. It now points you to README and is more of an
installation/build list of troubleshooting pointers.