Add a transfer target to build and test both 2.5+nw and 2.5 Lua
First step in moving up to using Lua 5 is removing the nw extensions
that were added to 2.5.

In order to provide a level of confidence of correctness I am adding
a new target called `tranfer' because we are transferring from 2.5+nw
to 2.5, and eventually up to 5. This target was initiall modelled on
careful.

The plan is to build a 2.5+nw and a 2.5 version with the programs in
tools. Then build them again using the 2.5+nw version. Then build them
one more time using the 2.5 version. If the 2.5+nw and 2.5 versions are
the same regardless of whether they were built with 2.5+nw or 2.5 then
both programs should be working the same and the conversion of the Lua
scripts was correct.

A quick look for echo commands will show how much of it works so far.