Using nocond to create two simultaneous versions of the lua scripts.
The prior version is generated by `nocond luanw' and the updated one
is generated by `nocond lua2'.

The root Makefile has a completed recipe for `trasfer' that builds
both versions of Noweb 3 with the Lua-2.5+nw library, and then builds
both versions of Noweb 3 with the Lua-2.5 library, and compares the
results.

With this commit, the comparison shows no differences so I believe
that the translation from `globmatch' and `case end' to `matchpattern'
and `if elseif end' is correct.