Update the developer documenation.
I was looking over the Lua2 branch and deciding
how similar or different the transition to Lua3
should be and noticed that the README was incorrect.

There are still several references to Lua-2.5+nw
in the xdoc folder. I have opted to leave those as
is for now because todo.nw already has a line item
acknowledging that all the text in xdoc needs a going
over.

I fully admit that this is not literate but have you
read the chunck of text that appears on page one of
the noweb weave?